fixing some of the cycles in glimmer-vm
