+++ packages/@glimmer/util/index.ts
@@ -1,3 +1,4 @@
+import { LOCAL_LOGGER, LOGGER } from './lib/local-logger';
 export * from './lib/array-utils';
 export { dict, isDict, isIndexable, StackImpl as Stack } from './lib/collections';
 export { beginTestSteps, endTestSteps, logStep, verifySteps } from './lib/debug-steps';
@@ -12,23 +13,7 @@ export * from './lib/string';
 
 export type FIXME<T, S extends string> = (T & S) | T;
 
-/**
- * This constant exists to make it easier to differentiate normal logs from
- * errant console.logs. LOCAL_LOGGER should only be used inside a
- * LOCAL_TRACE_LOGGING check.
- *
- * It does not alleviate the need to check LOCAL_TRACE_LOGGING, which is used
- * for stripping.
- */
-export const LOCAL_LOGGER = console;
-
-/**
- * This constant exists to make it easier to differentiate normal logs from
- * errant console.logs. LOGGER can be used outside of LOCAL_TRACE_LOGGING checks,
- * and is meant to be used in the rare situation where a console.* call is
- * actually appropriate.
- */
-export const LOGGER = console;
+export { LOCAL_LOGGER, LOGGER };
 
 export function assertNever(value: never, desc = 'unexpected unreachable branch'): never {
   LOGGER.log('unreachable', value);
